Factory Past Grid Setup and a bunch of restructuring things for it
