+++ Slider/Assets/Prefabs/Map/AreaSTiles/TimeTravelSTile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9cce9861f60df8048b1a9d380adda9fd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
