+++ Slider/Assets/Scripts/Sliders/Grid/TimeTravelSTile.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TimeTravelSTile : STile
+{
+    public override Vector3 calculatePosition(int x, int y)
+    {
+        return new Vector3(x / 3 * ((TimeTravelGrid)SGrid.current).gridOffset + x % 3 * STILE_WIDTH, y * STILE_WIDTH);
+    }
+
+    public override Vector3 calculateMovingPosition(float x, float y)
+    {
+        Vector3 newPos = STILE_WIDTH * new Vector3(x, y);
+
+        if (x >= 3)
+            newPos += new Vector3(((TimeTravelGrid)SGrid.current).gridOffset - 3 * STILE_WIDTH, 0);
+        return newPos;
+    }
+}
