+++ Slider/Assets/Scripts/Sliders/Grid/TimeTravelSTile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e985131fb1452549b503ef6ab9bcc7e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
